-Moved pages that needs authentication to WEB-INF folder -Created the javascript files separetade from the folder -Changed attribut Telephone from Client to double because Integer's limit -Implemented logout -Implemented forward of pages and verification if user is logged -Implemented menu according if user is logged or not -Cleaned the code(Got debug code out) -Cleaned the database -Tested the whole application
Pending:
-Les annees listees sont seulement 2015 et 2018, il faut ajouter une range a partir de 1950 a aujourd'hui

-Pendent le choix des pieces, si aucun choix de cote ou position n'est pas choisit, il insere comme memme et il va avoir une exception. On doit blocker l' ajout s' il ny a aucune choix ou incomplete

-Si il y a une piece qui n'a pas cote ou position, il faut comme memme envoyer l' information "na"  pour la cote ou position qui n' existe pas, le problemme arrive au moment qui le methode addPieceva verifier si la cote de la piece qui on ajoute est egal a cote de la piece qui est deja dans le " pannier" 

-Si un utilisateur avec mot de passe ou nom utilisateur pas bon essaye de faire login, j'ai demande de aller a la page d'erreur, le correct c'est de envoyer une message.

-Une fois qui on va au " pannier"  si on click sur faire commande, il devrai continuer la commande ancien, pas commencer une nouvelle. On doit verifier si la commande existe, s'elle existe, on doit " continuer"

+++ src/java/model/Client.java
@@ -13,11 +13,11 @@ public class Client {
     private int idClient;
     private String nomClient;
     private String adresse;
-    private int telephone;
+    private double telephone;
     private String usager;
     private String pass;
 
-    public Client(int idClient, String nomClient, String adresse, int telephone, String usager, String pass) {
+    public Client(int idClient, String nomClient, String adresse, double telephone, String usager, String pass) {
         this.idClient = idClient;
         this.nomClient = nomClient;
         this.adresse = adresse;
@@ -26,7 +26,7 @@ public class Client {
         this.pass = pass;
     }
 
-    public Client(int idClient, String nomClient, String adresse, int telephone) {
+    public Client(int idClient, String nomClient, String adresse, double telephone) {
         this.idClient = idClient;
         this.nomClient = nomClient;
         this.adresse = adresse;
@@ -59,7 +59,7 @@ public class Client {
         this.adresse = adresse;
     }
 
-    public int getTelephone() {
+    public double getTelephone() {
         return telephone;
     }
